Move Gemini streaming usage test into agent_test.go
Address review feedback: consolidate into the package's canonical test file.
